refactor: alias and auto-import were added, the ToDo component was added, the logic was transferred to ToDo from the App
