+++ src/App.jsx
@@ -1,0 +1,9 @@
+import ToDo from '@/components/ToDo/ToDo'
+
+const App = () => {
+  return (
+    <ToDo />
+  )
+}
+
+export default App
