+++ src/components/CircularTimer/CircularTimer.module.scss
