+++ src/icons/CheckmarkIcon/CheckmarkIcon.module.scss
