+++ src/icons/CloseIcon/CloseIcon.module.scss
