+++ src/icons/DarkThemeIcon/DarkThemeIcon.module.scss
