+++ src/icons/DeleteIcon/DeleteIcon.jsx
