+++ src/icons/DeleteIcon/DeleteIcon.module.scss
