+++ src/icons/EditIcon/EditIcon.jsx
