+++ src/icons/EditIcon/EditIcon.module.scss
