+++ src/icons/LightThemeIcon/LightThemeIcon.jsx
