+++ src/icons/LightThemeIcon/LightThemeIcon.module.scss
