+++ src/icons/SearchIcon/SearchIcon.module.scss
