+++ src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './components/App/App'
-import './styles/main.scss'
+import App from '@/App'
+import '@/styles/main.scss'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
